Move app into a container remove CRA crust add variable css and reset add adobe typekit link to index
