+++ public/index.html
@@ -24,6 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <link rel="stylesheet" href="https://use.typekit.net/qmg6jpo.css">
     <title>Soft Bodies</title>
   </head>
   <body>
