+++ src/config.js
@@ -34,14 +34,6 @@ export default {
             name: 'Leming Z/C',
             slug: 'leming-zc'
         }
-        
+
     ],
-    style: {
-        background: '#000000',
-        activeFill: '#e3f994',
-        blobFill: 'rgb(185,205,210)',
-        blobStroke: 'rgb(255,255,255)',
-        blobFont: '24px Helvetica Neue',
-        blobTextFill: 'rgb(255,255,255)'
-    }
 }
